+++ packages/renderer/src/Pages/normline/widgets/preprocessing/TheoryRow.svelte
@@ -6,7 +6,7 @@
     import QuickBrowser from '$components/QuickBrowser.svelte'
     import CustomSwitch from '$components/CustomSwitch.svelte'
     import computePy_func from '$src/Pages/general/computePy'
-
+    import BrowseTextfield from '$src/components/BrowseTextfield.svelte'
     /////////////////////////////////////////////////////////////////////////
 
     // export let normMethod
@@ -63,19 +63,33 @@
 />
 
 {#if $toggleRow}
-    <div class="align">
-        <button
-            class="button is-link"
-            on:click={() => {
-                showTheoryFiles = !showTheoryFiles
-            }}
-        >
-            Browse File</button
-        >
-        <CustomTextSwitch style="width:7em;" variant="outlined" bind:value={sigma} label="Sigma" step="0.5" />
-        <CustomTextSwitch style="width:7em" variant="outlined" bind:value={scale} label="Scale" step="0.001" max="1" />
-        <CustomSwitch style="margin: 0 1em;" bind:selected={onlyExpRange} label="Only Exp. Range" />
-        <CustomSwitch style="margin: 0 1em;" bind:selected={tkplot} label="Matplotlib" />
-        <button class="button is-link" on:click={plotData}>Replot</button>
-    </div>
+    <BrowseTextfield
+        class="two_col_browse box"
+        style="border: solid 1px;"
+        bind:value={theoryLocation}
+        label="Theory location"
+    />
+
+    {#if window.fs.isDirectory(theoryLocation)}
+        <div class="align">
+            <button
+                class="button is-warning"
+                on:click={() => {
+                    showTheoryFiles = !showTheoryFiles
+                }}>Show File</button
+            >
+            <CustomTextSwitch style="width:7em;" variant="outlined" bind:value={sigma} label="Sigma" step="0.5" />
+            <CustomTextSwitch
+                style="width:7em"
+                variant="outlined"
+                bind:value={scale}
+                label="Scale"
+                step="0.001"
+                max="1"
+            />
+            <CustomSwitch style="margin: 0 1em;" bind:selected={onlyExpRange} label="Only Exp. Range" />
+            <CustomSwitch style="margin: 0 1em;" bind:selected={tkplot} label="Matplotlib" />
+            <button class="button is-link" on:click={plotData}>Replot</button>
+        </div>
+    {/if}
 {/if}