+++ packages/renderer/src/components/QuickBrowser.svelte
@@ -12,11 +12,11 @@
     const dispatch = createEventDispatcher()
 </script>
 
-<svelte:window
+<!-- <svelte:window
     on:keydown={(e) => {
         if (e.keyCode === 27) active = false
     }}
-/>
+/> -->
 
 <QuickView bind:active bind:currentLocation {title}>
     <FileBrowser bind:currentLocation bind:fileChecked {filetype} on:markedFile />